Add Pull Request Template (#14)
# Pull Request

## Description

This pull request includes a small change to the
`.github/pull_request_template.md` file. The change adds a new template
for pull requests to ensure that a summary of changes, relevant
motivation, context, and dependencies are included.

*
[`.github/pull_request_template.md`](diffhunk://#diff-b2496e80299b8c3150b1944450bd81c622e04e13d15c411d291db0927d75fd6bR1-R7):
Added a new template section for pull requests to include a summary of
the change, issue fixed, motivation, context, and dependencies.

Fixes #1