Label Pull Request (#17)
# Pull Request

## Description

This pull request introduces a new GitHub Actions workflow for labeling
pull requests and automatically assigning size labels based on the
number of lines changed. It also includes a configuration file for the
labeler.

### New GitHub Actions Workflow:

*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfR1-R36):
Added a new workflow named "Pull Request Tasks" to label pull requests
and add size labels based on the number of lines changed. The workflow
triggers on pull request events and uses the `actions/labeler` and
`pascalgn/size-label-action` actions.

### Labeler Configuration:

*
[`.github/other-configurations/labeller.yml`](diffhunk://#diff-b490f987d9b055ce6e06e37d13763d61f6e0d5ee349b390bbb910f8d88ff36b9R1-R44):
Added a configuration file for the labeler to automatically assign
labels based on the types of files changed in a pull request.

Fixes #10