Add Dependency Review (#23)
# Pull Request

## Description

This pull request adds a new job to the GitHub Actions workflow to
perform a dependency review. The most important change includes the
addition of a `dependency-review` job in the
`.github/workflows/pull-request-tasks.yml` file.

Enhancements to GitHub Actions workflow:

*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfR37-R48):
Added a new job named `dependency-review` that runs on `ubuntu-latest`,
checks out the repository, and performs a dependency review using the
`actions/dependency-review-action@v4.5.0`.

Fixes #12