Regularly Clear Caches (#42)
# Pull Request

## Description

This pull request introduces a new workflow to clean GitHub Action
caches. The workflow is scheduled to run monthly and can also be
triggered manually.

*
[`.github/workflows/clean-caches.yml`](diffhunk://#diff-d0394e4336a74cdfc1d4cff05d056b893ac7ff922eacf4448e104a754f386b8dR1-R24):
Added a new workflow named "Clean Caches" that runs on a schedule
(monthly) and can be triggered manually. It includes permissions to read
contents and a job to clean GitHub Action caches using the `gh cache
delete` command.