Add CodeQL configuration (#45)
# Pull Request

## Description

This pull request includes changes to the CodeQL configuration and
workflow files to enhance security and quality checks. The most
important changes include the addition of a query filter to exclude
unpinned tags and the inclusion of a configuration file in the CodeQL
analysis step.

Improvements to CodeQL configuration:

*
[`.github/other-configurations/codeql-config.yml`](diffhunk://#diff-687a0af9316072cbdab79cdb046aac181231f8b5f3b5502bc796a9d949ce65beR1-R3):
Added a query filter to exclude the `actions/unpinned-tag` query.

Enhancements to CodeQL workflow:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R96):
Included the `config-file` parameter to reference the newly added CodeQL
configuration file.