Check Justfile Format (#49)
# Pull Request

## Description

This pull request includes updates to the GitHub Actions workflows in
the `.github/workflows/code-checks.yml` file. The changes focus on
adding a new job for checking the Justfile format and updating the
versions of several actions used in the workflow.

Updates to GitHub Actions workflows:

* Added a new job `check-justfile-format` to validate the format of the
Justfile.
* Updated the version of `astral-sh/setup-uv` action from `v5.2.2` to
`v5.3.0`.
* Updated the version of `github/codeql-action/upload-sarif` action from
`v3.28.8` to `v3.28.10`.
* Updated the version of `github/codeql-action/init` action from
`v3.28.8` to `v3.28.10`.
* Updated the version of `github/codeql-action/analyze` action from
`v3.28.8` to `v3.28.10`.

Fixes #50