Update Labels (#56)
# Pull Request

## Description

This pull request includes updates to the label configuration files to
better categorize and manage pull requests based on the types of changes
they introduce. The key changes are the addition of new labels and the
reorganization of existing ones.

Label configuration updates:

*
[`.github/other-configurations/labeller.yml`](diffhunk://#diff-b490f987d9b055ce6e06e37d13763d61f6e0d5ee349b390bbb910f8d88ff36b9R2-R20):
Added new label categories for `dependencies` and `configurations`, and
reorganized existing labels under new categories such as `python`,
`typescript`, and `just`.
[[1]](diffhunk://#diff-b490f987d9b055ce6e06e37d13763d61f6e0d5ee349b390bbb910f8d88ff36b9R2-R20)
[[2]](diffhunk://#diff-b490f987d9b055ce6e06e37d13763d61f6e0d5ee349b390bbb910f8d88ff36b9L16-R56)
*
[`.github/other-configurations/labels.yml`](diffhunk://#diff-c1b32711b0f3da3f5c7007cfaa4d9e94ae5430fea1f4e1256f6210bc6988553dR5-R7):
Added new labels for `configurations`, `python`, and `typescript`, and
reintroduced the `just` label. Removed redundant labels such as `docker`
and `analyser`.
[[1]](diffhunk://#diff-c1b32711b0f3da3f5c7007cfaa4d9e94ae5430fea1f4e1256f6210bc6988553dR5-R7)
[[2]](diffhunk://#diff-c1b32711b0f3da3f5c7007cfaa4d9e94ae5430fea1f4e1256f6210bc6988553dL36-L51)