Add GitHub Action Step Name (#77)
# Pull Request

## Description

This pull request includes minor updates to configuration files for
GitHub Actions and Lefthook. The changes improve clarity and ensure
compatibility with newer versions.

Updates to GitHub Actions:

*
[`.github/workflows/sync-labels.yml`](diffhunk://#diff-a877ed9f27d115d95934fd904f2475dcec6ce4125da686dd5b3c75a696fff1c6L25-R26):
Added a descriptive name ("Sync labels") to the
`micnncim/action-label-syncer` step for better readability.

Updates to Lefthook configuration:

*
[`lefthook.yml`](diffhunk://#diff-ad6a01e589b8b1b214ca310dbb8d2e4314f6c612b921050c73c97455de43884dL2-R2):
Updated the `min_version` requirement from `1.11.6` to `1.11.11` to
ensure compatibility with the latest Lefthook features and fixes.