Update Dependabot Schedule (#87)
# Pull Request

## Description

This pull request updates the Dependabot configuration to use a
cron-based schedule instead of a daily interval. This change provides
more precise control over when dependency updates are checked.

*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28L8-R9):
Changed the schedule configuration from a daily interval with a specific
time and timezone to a cronjob format (`30 7 * * *`) for better
flexibility and precision.