Add Zizmor Configuration (#92)
# Pull Request

## Description

This pull request updates the `lefthook-validate` configuration in the
`Justfile` to enhance the strictness of the `zizmor` checks by
introducing pedantic mode and specifying a persona.

Enhancements to `zizmor-check`:

*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cL47-R47):
Modified the `zizmor-check` command to include the `--pedantic` flag and
set the `--persona` option to `pedantic`, ensuring stricter validation
during checks.