Remove `.editorconfig-checker.json` (#106)
# Pull Request

## Description

This pull request removes the configuration file for the EditorConfig
Checker and updates the GitHub Actions workflow to reflect this change.

### Removal of EditorConfig Checker:

*
[`.github/super-linter-configurations/.editorconfig-checker.json`](diffhunk://#diff-926aaab00380cdd79b8c5069e4eb6d6a7b0a730b2cfd8254e271be7844a0bd10L1-L10):
Deleted the entire configuration file, as it is no longer needed.

### Workflow updates:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L35):
Removed the `EDITORCONFIG_FILE_NAME` environment variable from the
GitHub Actions workflow, as the EditorConfig Checker is no longer in
use.