Migrate Clear Caches to Common Clear Caches (#108)
# Pull Request

## Description

This pull request updates the `.github/workflows/clean-caches.yml` file
to simplify and standardize the cache-cleaning workflow by switching to
a reusable workflow and adjusting permissions.

### Workflow simplification and standardization:

* Replaced the inline `clean-caches` job with a reusable workflow from
`JackPlowman/reusable-workflows` to streamline the maintenance and
execution of the cache-cleaning process.
* Adjusted permissions from `contents: read` to an empty object (`{}`)
at the workflow level and redefined `contents: read` within the reusable
workflow's job, aligning with best practices for permissions management.