Sort Outputs of Commands (#112)
# Pull Request

## Description

This pull request updates the `README.md` file to improve the clarity
and functionality of example commands for listing repositories. The most
notable change is the addition of a sorting step to the output of the
commands.

Documentation improvements:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L30-R35):
Added a `| sort` step to the example commands for listing repository
names and names with owners, ensuring that the output is sorted
alphabetically.