Add project status (#137)
# Pull Request

## Description

This pull request updates the `README.md` file to enhance its structure
and readability by adding a maintenance status badge and a Table of
Contents section.

Documentation improvements:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R3-R8):
Added a maintenance status badge at the top of the file to indicate
project status.
*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R3-R8):
Introduced a Table of Contents section to improve navigation within the
document.