Add ActionLint Git Hook (#151)
# Pull Request

## Description

This pull request introduces a new `actionlint` check to the project,
ensuring that GitHub Actions workflows are validated for correctness.
The changes include updates to the `Justfile` and `lefthook.yml` to
integrate this new check into the development workflow.

### Additions for `actionlint` integration:

* **`Justfile`:** Added a new `actionlint-check` recipe to run
`actionlint` for validating GitHub Actions workflows.
* **`lefthook.yml`:** Integrated the `actionlint-check` recipe into the
`pre-commit` hooks, allowing developers to automatically run
`actionlint` as part of the pre-commit validation process.