Add GitHub Workflow Concurrency (#194)
# Pull Request

## Description

This pull request updates several GitHub Actions workflow files to add
concurrency controls. The main improvement is the introduction of a
mechanism to prevent multiple runs of the same workflow for the same
pull request or branch, ensuring that only the latest run is active and
prior runs are cancelled if a new one is triggered.

Concurrency management added to workflows:

* Added a `concurrency` block to `.github/workflows/clean-caches.yml` to
group runs by workflow and pull request or ref, cancelling in-progress
runs when a new one starts.
* Added a `concurrency` block to `.github/workflows/code-checks.yml`
with similar grouping and cancellation logic.
* Added a `concurrency` block to
`.github/workflows/pull-request-tasks.yml` to manage parallel workflow
executions for pull requests.
* Added a `concurrency` block to `.github/workflows/sync-labels.yml` to
ensure only one workflow run per group is active.