Reorder GitHub Jobs (#207)
# Pull Request

## Description

This pull request reorganizes the order of jobs in the
`.github/workflows/code-checks.yml` workflow file. The main change is
moving the definition of the `common-code-checks` job to appear after
the `codeql-checks` job. There are no functional changes—just a
restructuring for clarity or consistency.

- Workflow job reordering:
* Moved the `common-code-checks` job definition to follow the
`codeql-checks` job within `.github/workflows/code-checks.yml` for
improved organization.
[[1]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L22-L32)
[[2]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R33-R43)